Cleanup of tests
Organize tests in modules and reenable npk integration tests.
