+++ northstar-tests/tests/tests.rs
@@ -91,16 +91,14 @@ test!(mount_umount, {
     runtime.install(&TEST_CONTAINER_NPK, "test-0").await?;
     runtime.install(&TEST_RESOURCE_NPK, "test-0").await?;
 
-    for _ in 0..10u32 {
-        let mut containers = runtime.containers().await?;
-        runtime
-            .mount(containers.drain(..).map(|c| c.container))
-            .await?;
+    let mut containers = runtime.containers().await?;
+    runtime
+        .mount(containers.drain(..).map(|c| c.container))
+        .await?;
 
-        let containers = &mut runtime.containers().await?;
-        for c in containers.iter().filter(|c| c.mounted) {
-            runtime.umount(c.container.clone()).await?;
-        }
+    let containers = &mut runtime.containers().await?;
+    for c in containers.iter().filter(|c| c.mounted) {
+        runtime.umount(c.container.clone()).await?;
     }
 
     runtime.shutdown().await
@@ -387,7 +385,7 @@ test!(seccomp_allowed_syscall_with_prohibited_arg, {
 // Iterate all exit codes in the u8 range
 test!(exitcodes, {
     let mut runtime = Northstar::launch_install_test_container().await?;
-    for c in 0..=255 {
+    for c in &[0, 1, 10, 127, 128, 255] {
         runtime
             .start_with_args(TEST_CONTAINER, ["exit".to_string(), c.to_string()])
             .await?;
@@ -395,7 +393,7 @@ test!(exitcodes, {
             Notification::Exit {
                 status: ExitStatus::Exit { code },
                 ..
-            } if code == &c => true,
+            } if code == c => true,
             _ => false,
         };
         runtime.assume_notification(n, 5).await?;
@@ -457,142 +455,3 @@ test!(check_api_version_on_connect, {
 
     runtime.shutdown().await
 });
-
-mod example {
-    use super::*;
-
-    // Start crashing example
-    test!(crashing, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_CRASHING_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_CRASHING).await?;
-        runtime
-            .assume_notification(
-                |n| {
-                    matches!(
-                        n,
-                        Notification::Exit {
-                            status: ExitStatus::Signalled { signal: 6 },
-                            ..
-                        }
-                    )
-                },
-                20,
-            )
-            .await?;
-        runtime.shutdown().await
-    });
-
-    // Start console example
-    test!(console, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_CONSOLE_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_CONSOLE).await?;
-        // The console example stop itself - so wait for it...
-        assume("Client console:0.0.1 connected", 5).await?;
-        assume("Killing console:0.0.1 with SIGTERM", 5).await?;
-        runtime.shutdown().await
-    });
-
-    // Start cpueater example and assume log message
-    test!(cpueater, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_CPUEATER_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_CPUEATER).await?;
-        assume("Eating CPU", 5).await?;
-
-        runtime.stop(EXAMPLE_CPUEATER, 10).await?;
-        runtime.shutdown().await
-    });
-
-    // Start hello-ferris example
-    test!(hello_ferris, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_FERRIS_NPK, "test-0").await?;
-        runtime
-            .install(&EXAMPLE_MESSAGE_0_0_1_NPK, "test-0")
-            .await?;
-        runtime.install(&EXAMPLE_HELLO_FERRIS_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_HELLO_FERRIS).await?;
-        assume("Hello once more from 0.0.1!", 5).await?;
-        // The hello-ferris example terminates after printing something.
-        // Wait for the notification that it stopped, otherwise the runtime
-        // will try to shutdown the application which is already exited.
-        runtime
-            .assume_notification(
-                |n| {
-                    matches!(
-                        n,
-                        Notification::Exit {
-                            status: ExitStatus::Exit { code: 0 },
-                            ..
-                        }
-                    )
-                },
-                15,
-            )
-            .await?;
-
-        runtime.shutdown().await
-    });
-
-    // Start hello-resource example
-    test!(hello_resource, {
-        let mut runtime = Northstar::launch().await?;
-        runtime
-            .install(&EXAMPLE_MESSAGE_0_0_2_NPK, "test-0")
-            .await?;
-        runtime
-            .install(&EXAMPLE_HELLO_RESOURCE_NPK, "test-0")
-            .await?;
-        runtime.start(EXAMPLE_HELLO_RESOURCE).await?;
-        assume(
-            "0: Content of /message/hello: Hello once more from v0.0.2!",
-            5,
-        )
-        .await?;
-        assume(
-            "1: Content of /message/hello: Hello once more from v0.0.2!",
-            5,
-        )
-        .await?;
-        runtime.shutdown().await
-    });
-
-    // Start inspect example
-    test!(inspect, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_INSPECT_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_INSPECT).await?;
-        runtime.stop(EXAMPLE_INSPECT, 5).await?;
-        // TODO
-        runtime.shutdown().await
-    });
-
-    // Start memeater example
-    // test!(memeater, {
-    //     let mut runtime = Northstar::launch().await?;
-    //     runtime.install(&EXAMPLE_MEMEATER_NPK, "test-0").await?;
-    //     runtime.start(EXAMPLE_MEMEATER).await?;
-    //     assume("Process memeater:0.0.1 is out of memory", 20).await?;
-    //     runtime.shutdown().await
-    // });
-
-    // Start persistence example and check output
-    test!(persistence, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_PERSISTENCE_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_PERSISTENCE).await?;
-        assume("Writing Hello! to /data/file", 5).await?;
-        assume("Content of /data/file: Hello!", 5).await?;
-        runtime.shutdown().await
-    });
-
-    // Start seccomp example
-    test!(seccomp, {
-        let mut runtime = Northstar::launch().await?;
-        runtime.install(&EXAMPLE_SECCOMP_NPK, "test-0").await?;
-        runtime.start(EXAMPLE_SECCOMP).await?;
-        runtime.shutdown().await
-    });
-}
